+++ RenderSystems/Direct3D9/src/OgreD3D9Device.cpp
@@ -604,7 +604,7 @@ namespace Ogre
 			hr = mpDevice->SetTexture(stage, NULL);
 			if( hr != S_OK )
 			{
-				String str = "Unable to disable texture '" + StringConverter::toString(stage) + "' in D3D9";
+				String str = "Unable to disable texture '" + StringConverter::toString((unsigned int)stage) + "' in D3D9";
 				OGRE_EXCEPT(Exception::ERR_RENDERINGAPI_ERROR, str, "D3D9Device::clearDeviceStreams" );
 			}
 		
@@ -615,7 +615,7 @@ namespace Ogre
 				hr = mpDevice->SetTextureStageState(static_cast<DWORD>(stage), D3DTSS_COLOROP, D3DTOP_DISABLE);
 				if( hr != S_OK )
 				{
-					String str = "Unable to disable texture '" + StringConverter::toString(stage) + "' in D3D9";
+					String str = "Unable to disable texture '" + StringConverter::toString((unsigned)stage) + "' in D3D9";
 					OGRE_EXCEPT(Exception::ERR_RENDERINGAPI_ERROR, str, "D3D9Device::clearDeviceStreams" );
 				}
 			}
